backend/requirements.txt
fastapi
uvicorn
pydantic
pandas
python-multipart
google-generativeai
python-dotenv
